requirements.txt
aiogram~=3.15
environs~=11.0.0
cachetools~=5.3.0

# Chess analysis
python-chess~=1.999
stockfish~=3.28.0

# AI integration
google-genai~=1.47.0

# Database
aiosqlite~=0.20.0

# Image processing
pillow~=10.4.0
